test: extract file tool white-box suite
